+++ src/Colour.h
@@ -5,15 +5,15 @@
 
 class Colour {
 public:
-    float red;
-    float green;
-    float blue;
+    double red;
+    double green;
+    double blue;
 
-    Colour(float red, float green, float blue);
+    Colour(double red, double green, double blue);
     Colour();
     Colour add(const Colour& other) const;
     Colour subtract(const Colour& other) const;
-    Colour scale(float s) const;
+    Colour scale(double s) const;
     Colour multiply(const Colour& other) const;
     bool equals(const Colour& other) const {
         return almostEqual(red, other.red) && almostEqual(green, other.green) && almostEqual(blue, other.blue);
